Move connection functions to connectionsModel.js
